+++ packages/ui/__tests__/components/Layout.test.tsx
@@ -9,12 +9,11 @@
 // WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
 // PARTICULAR PURPOSE. See the GNU General Public License for more details.
 
-import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { withChakraTheme } from '../test-utilities';
-import { Layout, LayoutProps } from '../../src/components';
+import Layout from '../../src/components/Layout';
 
-const Component = withChakraTheme<LayoutProps>(Layout);
+const Component = withChakraTheme(Layout);
 const defaultProps = {
     headerLinks: [
         {
